[MIN-475] Move exchange_rate calculation from liquidity_pools to controller
