+++ pallets/risk-manager/src/lib.rs
@@ -538,7 +538,7 @@ impl<T: Config> Pallet<T> {
 
 		// Get an array of collateral pools for the borrower.
 		// The array is sorted in descending order by the number of wrapped tokens in USD.
-		let collateral_pools = <LiquidityPools<T>>::get_is_collateral_pools(&borrower)?;
+		let collateral_pools = <T as Config>::ControllerManager::get_is_collateral_pools(&borrower)?;
 
 		// Collect seized pools.
 		let mut seized_pools: Vec<CurrencyId> = Vec::new();
@@ -558,7 +558,7 @@ impl<T: Config> Pallet<T> {
 				// seize_tokens = seize_amount / (price_collateral * exchange_rate)
 				let price_collateral =
 					T::PriceSource::get_underlying_price(collateral_pool_id).ok_or(Error::<T>::InvalidFeedPrice)?;
-				let exchange_rate = <LiquidityPools<T>>::get_exchange_rate(collateral_pool_id)?;
+				let exchange_rate = <T as Config>::ControllerManager::get_exchange_rate(collateral_pool_id)?;
 				let seize_tokens = Rate::from_inner(seize_amount)
 					.checked_div(
 						&price_collateral
@@ -577,7 +577,7 @@ impl<T: Config> Pallet<T> {
 					Ordering::Less => {
 						// seize_underlying = balance_wrapped_token * exchange_rate
 						let seize_underlying =
-							<LiquidityPools<T>>::convert_from_wrapped(wrapped_id, balance_wrapped_token)?;
+							<T as Config>::ControllerManager::convert_from_wrapped(wrapped_id, balance_wrapped_token)?;
 						T::MultiCurrency::withdraw(wrapped_id, &borrower, balance_wrapped_token)?;
 						// seize_amount = seize_amount - (seize_underlying * price_collateral)
 						seize_amount -= Rate::from_inner(seize_underlying)
@@ -589,7 +589,8 @@ impl<T: Config> Pallet<T> {
 					// Enough collateral wrapped tokens. Transfer all seize_tokens to liquidation_pool.
 					_ => {
 						// seize_underlying = seize_tokens * exchange_rate
-						let seize_underlying = <LiquidityPools<T>>::convert_from_wrapped(wrapped_id, seize_tokens)?;
+						let seize_underlying =
+							<T as Config>::ControllerManager::convert_from_wrapped(wrapped_id, seize_tokens)?;
 						T::MultiCurrency::withdraw(wrapped_id, &borrower, seize_tokens)?;
 						// seize_amount = 0, since all seize_tokens have already been withdrawn
 						seize_amount = Balance::zero();
